feat: Add documentation for API and Ajax communication
This commit adds documentation explaining the concepts of asynchronous communication, specifically focusing on API communication and Ajax. It also provides an overview of the Fetch API and Axios library for performing asynchronous communication in React projects. The documentation includes examples of using Fetch API and Axios in a React + TypeScript application.